feat(modal): remove data-attributes, return BsModalRef, update docs
