+++ demo/src/app/components/+modal/modal-section.component.ts
@@ -77,10 +77,9 @@ let titleDoc = require('html-loader!markdown-loader!./docs/title.md');
   <h2 routerLink="." fragment="service" id="service">Modal service</h2>
   <p>Open a modal from service</p>
   <p>To be able to open modals from service, inject BsModalService, ElementRef, ViewContainerRef and Renderer to your constructor. <br>
-  Then prepare modal service like this <code>this.modalService = this.modalService.create(this.element, this.vcRef, this.renderer)</code>. <br>
-    Finally, call <code>.show()</code> method of modal service. Pass a TemplateRef or a component as a first argument and config as a second (optionally). <br>
+  Then prepare modal service like this <code>this.modalService = this.modalService.create(this.element, this.vcRef, this.renderer)</code>. <br><br>
+    Finally, call <code>.show()</code> method of modal service. Pass a TemplateRef or a component as a first argument and config as a second (optionally). If you're passing a component, add it to <code>entryComponents</code> of your <code>NgModule</code><br><br>
     <code>.show()</code> method returns an object with <code>.hide()</code> method and <code>content</code> property where you'll find a component which you've passed to service <br>
-    You can also define an area of template which should trigger hiding of modal by clicking on it. Just add <code>data-bsmodal-close="true"</code> <br>
     If you passed a component to <code>.show()</code> you can get access to opened modal by injecting BsModalRef. See example for more info
   </p>
   <ng-sample-box [ts]="demos.service.component" [html]="demos.service.html">
