+++ src/modal/modal-options.class.ts
@@ -28,6 +28,10 @@ export class BsModalRef {
    * Hides the modal
    */
   hide(): void {}
+  /**
+   * Reference to a component inside the modal. Null if modal's been created with TemplateRef
+   */
+  content?: Object | null;
 }
 
 export const modalConfigDefaults:ModalOptions = {
